fix: refactor to work together
